Moving the OPLSAA.sb scanner from a separate file to FileUtilities.
